+++ js/templates/BookmarkDetail.html
@@ -3,6 +3,7 @@
 	<span class="message message-saving"><span class="icon-loading"></span></span>
 	<span class="message message-saved"><span class="icon-checkmark"></span> <%- t('bookmarks', 'Saved') %></span>
 </div>
+<div class="preview"></div>
 <h1 data-attribute="title"><%- title %></h1>
 <h3>Link</h3>
 <h2 data-attribute="url">
